Altered faulty migration file

+++ db/migrate/20230329084610_drop_questions_time.rb
@@ -1,5 +1,9 @@
-class DropQuestionsTime < ActiveRecord::Migration[7.0]
-  def change
-    drop_table :questions_times
+class DropQuestionsTime < ActiveRecord::Migration[7.1]
+ def up
+    drop_table :questions_times if table_exists?(:questions_times)
+ end
+
+ def down
+    raise ActiveRecord::IrreversibleMigration
   end
 end
